move tests from match to where
